requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
speechrecognition==3.10.0
pyttsx3==2.90
openai-whisper==20231117
pyaudio==0.2.11
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
